--- src/pages/Interest.tsx
@@ -1,5 +1,0 @@
-const Interest = () => {
-  return <div>관심사 페이지입니다.</div>;
-};
-
-export default Interest;
